BFS based optimal cuckoo layout.
This does as much search as is necessary during BFS to ensure we're maximizing the number of keys that are in their first location.

This helps a reasonable amount. Up to 75% it's better than aligned but worse than unaligned. At about 75%-87.5% it's about a match for unaligned. Above 87.5% it starts beating unaligned.

It's very slow at insertions though.